+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## HEAD
+## v0.2.0 (2019-11-02)
 
 * Add compatibility with calendars other than `Calendar.ISO`.